requirements.txt
Django==1.9.6
Pillow==3.2.0
argparse==1.2.1
django-grappelli==2.8.1
djangorestframework==3.3.3
wsgiref==0.1.2
